infra: update loopback mac address
If the MAC address of a `gr-loop*` kernel interface changes,
the interface will not receive any packet, because the loopback
control function always set the src and dst MAC address to the
value got in the `iface_loopback_init` function.

The MAC address of tun interfaces can be automatically changed by
systemd-networkd logic. This happens in Github CI, as the default
runner is based on Ubuntu.

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>